Set file upload response timeout to 60 seconds
Seems like Google Cloud Storage takes more than a few seconds to provide
a response after receiving a large uploaded file. The default of five
seconds was not enough, and hackney was timing out. The value is now set
to 60 seconds, which should be sufficient, and avoids blocking forever.

rebar3 ct passes