Raise receive timeout to 5 minutes
Everything had been working fine for a while, but then this one small file
(~30MB) was repeatedly failing with a timeout error. Likely coming from
hackney due to the `recv_timeout` surpassing the 1 minute value. How this
could happen is a mystery, but try raising the limit to 5 minutes and see if
the error goes away.

rebar3 ct passes